Local branch amd-gfx d2b14a246377 Merged main:04f61fb73dc6 into amd-gfx:cc1f635ba5ab
Remote branch main 7a74869a7221 Fix mlir/lib/Bindings/Python/IRTypes.cpp for Float8E4M3B11FNUZType

+++ llvm/lib/ExecutionEngine/JITLink/aarch32.cpp
@@ -157,8 +157,9 @@ struct ThumbRelocation {
 Error makeUnexpectedOpcodeError(const LinkGraph &G, const ThumbRelocation &R,
                                 Edge::Kind Kind) {
   return make_error<JITLinkError>(
-      formatv("Invalid opcode [ 0x{0:x4}, 0x{1:x4} ] for relocation: {2}", R.Hi,
-              R.Lo, G.getEdgeKindName(Kind)));
+      formatv("Invalid opcode [ 0x{0:x4}, 0x{1:x4} ] for relocation: {2}",
+              static_cast<uint16_t>(R.Hi), static_cast<uint16_t>(R.Lo),
+              G.getEdgeKindName(Kind)));
 }
 
 template <EdgeKind_aarch32 Kind> bool checkOpcode(const ThumbRelocation &R) {
